+++ R/print.dispRity.R
@@ -266,7 +266,13 @@ print.dispRity <- function(x, all = FALSE, ...) {
                 cat(" ---- dispRity object ---- \n")
                 dims <- dim(x$matrix[[1]])
                 n_matrices <- length(x$matrix)
-                cat(paste0("Contains ", ifelse(n_matrices > 1, paste0(n_matrices, " matrices "), "a matrix "), dims[1], "x", dims[2], "."))
+                cat(paste0("Contains ", ifelse(n_matrices > 1, paste0(n_matrices, " matrices "), "a matrix "), dims[1], "x", dims[2]))
+                ## And trees?
+                if(!is.null(x$tree[[1]])) {
+                    n_trees <- length(x$tree)
+                    cat(paste0(" and ", ifelse(n_trees > 1, paste0(n_trees, " trees"), "a tree")))
+                }
+                cat(".")
             } else {
                 cat("Empty dispRity object.\n")
             }